Move the overlap functions to a more logical place
